LayoutTests/imported/w3c/web-platform-tests/intersection-observer/edge-inclusive-intersection-expected.txt


PASS IntersectionObserver should detect and report edge-adjacent and zero-area intersections. 
FAIL First rAF. assert_equals: entries[0].boundingClientRect.left expected 8 but got 0
FAIL Set transform=translateY(200px) on target. assert_equals: entries.length expected 2 but got 1
FAIL Set transform=translateY(201px) on target. assert_equals: entries.length expected 3 but got 1
FAIL Set transform=translateY(185px) on target. assert_equals: entries.length expected 4 but got 1
